REFACTOR : coworking 패키지 추가 & 공통 메서드 분리
- CoworkingSupport 추가
