+++ src/main/java/com/dikkak/controller/coworking/CoworkingController.java
@@ -1,0 +1,63 @@
+package com.dikkak.controller.coworking;
+
+import com.dikkak.common.BaseException;
+import com.dikkak.config.UserPrincipal;
+import com.dikkak.dto.coworking.GetChattingRes;
+import com.dikkak.dto.coworking.GetTaskRes;
+import com.dikkak.dto.coworking.Message;
+import com.dikkak.service.coworking.CoworkingService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+import static com.dikkak.common.ResponseMessage.INVALID_ACCESS_TOKEN;
+import static com.dikkak.common.ResponseMessage.UNAUTHORIZED_REQUEST;
+
+@RestController
+@RequestMapping("/coworking")
+@RequiredArgsConstructor
+public class CoworkingController {
+
+    private final CoworkingService coworkingService;
+    private final CoworkingSupport coworkingSupport;
+
+    /**
+     * 외주작업실의 채팅 목록 조회
+     * @param principal 회원 id, 타입
+     * @param coworkingId 외주작업실 id
+     */
+    @GetMapping("/chat")
+    public List<Message<GetChattingRes>> getChatList(
+            @AuthenticationPrincipal UserPrincipal principal,
+            @RequestParam Long coworkingId) throws BaseException {
+
+        if(principal == null) throw new BaseException(INVALID_ACCESS_TOKEN);
+        if(!coworkingSupport.checkUser(principal, coworkingId)) throw new BaseException(UNAUTHORIZED_REQUEST);
+
+        return coworkingService.getMessageList(coworkingId);
+    }
+
+    /**
+     * 외주작업실 task 조회
+     * @param principal 회원 id, 타입
+     * @param coworkingId 외주작업실 id
+     */
+    @GetMapping("/task")
+    public List<Message<GetTaskRes>> getTask(@AuthenticationPrincipal UserPrincipal principal,
+                                             @RequestParam Long coworkingId) throws BaseException {
+        if(principal == null) throw new BaseException(INVALID_ACCESS_TOKEN);
+        if(!coworkingSupport.checkUser(principal, coworkingId)) throw new BaseException(UNAUTHORIZED_REQUEST);
+
+        return coworkingService.getTaskList(coworkingId);
+    }
+
+
+
+
+
+}
